Refactor and cleanup analysis routine initialization
