+++ src/util/analysis/mod.rs
@@ -1,4 +1,5 @@
 use crate::scheduler::*;
+use crate::util::statistics::stats::Stats;
 use crate::vm::VMBinding;
 use crate::MMTK;
 use std::sync::{Arc, Mutex};
@@ -6,6 +7,10 @@ use std::sync::{Arc, Mutex};
 pub mod gc_count;
 pub mod obj_num;
 pub mod obj_size;
+
+use self::gc_count::GcCounter;
+use self::obj_num::ObjectCounter;
+use self::obj_size::PerSizeClassObjectCounter;
 
 /**
  * This trait exposes hooks for developers to implement their own analysis routines.
@@ -45,6 +50,19 @@ impl<VM: VMBinding> AnalysisManager<VM> {
         AnalysisManager {
             routines: Mutex::new(vec![]),
         }
+    }
+
+    // Initializing all routines. If you want to add a new routine, here is the place
+    // to do so
+    pub fn initialize_routines(&mut self, stats: &Stats) {
+        let ctr = stats.new_event_counter("obj.num", true, true);
+        let gc_ctr = stats.new_event_counter("gc.num", true, true);
+        let obj_num = Arc::new(Mutex::new(ObjectCounter::new(true, ctr)));
+        let gc_count = Arc::new(Mutex::new(GcCounter::new(true, gc_ctr)));
+        let obj_size = Arc::new(Mutex::new(PerSizeClassObjectCounter::new(true)));
+        self.add_analysis_routine(obj_num);
+        self.add_analysis_routine(gc_count);
+        self.add_analysis_routine(obj_size);
     }
 
     pub fn add_analysis_routine(&mut self, routine: Arc<Mutex<dyn RtAnalysis<VM> + Send>>) {
